py/kubeflow/testing/cloudprovider/aws/requirements.txt
# You should use python 3.8 and install this requirements.txt

attrs==20.3.0
boto3==1.17.17
botocore==1.20.17
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
fire==0.4.0
google-auth==1.27.0
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
junit-xml==1.9
kubernetes==12.0.1
mock==4.0.3
oauthlib==3.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
s3transfer==0.3.4
six==1.15.0
termcolor==1.1.0
toml==0.10.2
urllib3==1.26.3
websocket-client==0.57.0